test: Move custom Jest matchers to separate file (#1019)
* test: Move toMatchTaskWithDescription() to separate file

To allow for re-use.

* test: Add comment saying how to use extracted code

* test: Move toMatchTaskWithPath() & toBeValid() to separate file

* test: Combine 2 'namespace jest' blocks in to 1

* test: Make function order in file consistent

The functions are now implemented in the same order that they
are used in the 'namespace jest' block at the top of the file.

* test: Clarify a comment

Now that custom matchers are available.

* test: Move custom matchers to tests/CustomMatchers/
