+++ tests/Query/Filter/PathField.test.ts
@@ -2,6 +2,10 @@ import { PathField } from '../../../src/Query/Filter/PathField';
 import type { FilterOrErrorMessage } from '../../../src/Query/Filter/Filter';
 import { TaskBuilder } from '../../TestingTools/TaskBuilder';
 import { testFilter } from '../../TestingTools/FilterTestHelpers';
+import {
+    toBeValid,
+    toMatchTaskWithPath,
+} from '../../CustomMatchers/CustomMatchersForFilters';
 
 function testTaskFilterForTaskWithPath(
     filter: FilterOrErrorMessage,
@@ -10,69 +14,6 @@ function testTaskFilterForTaskWithPath(
 ) {
     const builder = new TaskBuilder();
     testFilter(filter, builder.path(path), expected);
-}
-
-declare global {
-    namespace jest {
-        interface Matchers<R> {
-            toMatchTaskWithPath(path: string): R;
-            toBeValid(): R;
-        }
-
-        interface Expect {
-            toMatchTaskWithPath(path: string): any;
-            toBeValid(): any;
-        }
-
-        interface InverseAsymmetricMatchers {
-            toMatchTaskWithPath(path: string): any;
-            toBeValid(): any;
-        }
-    }
-}
-
-export function toBeValid(filter: FilterOrErrorMessage) {
-    if (filter.filter === undefined) {
-        return {
-            message: () =>
-                'unexpected null filter: check your instruction matches your filter class',
-            pass: false,
-        };
-    }
-
-    if (filter.error !== undefined) {
-        return {
-            message: () =>
-                'unexpected error message in filter: check your instruction matches your filter class',
-            pass: false,
-        };
-    }
-
-    return {
-        message: () => 'filter is unexpectedly valid',
-        pass: true,
-    };
-}
-
-export function toMatchTaskWithPath(
-    filter: FilterOrErrorMessage,
-    path: string,
-) {
-    const builder = new TaskBuilder();
-    const task = builder.path(path).build();
-
-    const matches = filter.filter!(task);
-    if (!matches) {
-        return {
-            message: () => `unexpected failure to match task: ${path}`,
-            pass: false,
-        };
-    }
-
-    return {
-        message: () => `filter should not have matched task: ${path}`,
-        pass: true,
-    };
 }
 
 expect.extend({
